+++ src/de/uni/freiburg/iig/telematik/sepia/petrinet/PNTimeContext.java
@@ -2,8 +2,8 @@ package de.uni.freiburg.iig.telematik.sepia.petrinet;
 
 public interface PNTimeContext {
 	
-	public long getDelayPT(String placeName, String transitionName);
+	public double getDelayPT(String placeName, String transitionName);
 	
-	public long getDelayTP(String transitionName, String placeName);
+	public double getDelayTP(String transitionName, String placeName);
 
 }
